Change Tempfile behavior to create clones in same directory to prevent issues :tm: (#2)
* Change Tempfile behavior to create clones in same directory to prevent Issues :tm:

Some Minitest cops (and maybe others), as well as config in .rubocop.yml
may not like us placing file clones in some obscure directory outside or
inside the project. Instead we create a hidden file in the same
directory as the file we are working on. This prevents confusing
[p]atch/[a]pply bugs that seeming work correctly when running rubocop
directly but not inside our tool as we previously placed the clone in a
different directory.

* Test